PRUEBA/PRUEBA/requirements.txt
Flask==3.0.2
Flask-Limiter==3.5.0
pyOpenSSL==24.0.0
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.5
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0 